requirements.txt
asn1crypto==0.24.0
awscli==1.15.51
boto==2.48.0
boto3==1.7.50
botocore==1.10.50
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudinary==1.12.0
Collectfast==0.2.3
colorama==0.3.9
cryptography==2.2.2
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-ajax-selects==1.7.0
django-angular==2.0.3
django-bcp==0.2.2
django-bootstrap4==0.0.6
django-currentuser==0.2.2
django-filter==1.1.0
django-heroku==0.3.1
django-localflavor==2.0
django-localflavor-br==1.0
django-postgres-copy==2.3.2
django-redis-cache==1.6.5
django-storages==1.6.6
django-storages-redux==1.3
django-widget-tweaks==1.4.2
docutils==0.14
dropbox==8.9.0
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
jmespath==0.9.3
mock==2.0.0
oauthlib==2.1.0
pbr==4.0.4
pdfrw==0.4
Pillow==5.1.0
psycopg2==2.7.5
pyasn1==0.4.3
pycparser==2.18
PyJWT==1.6.4
Pyphen==0.9.4
python-dateutil==2.7.3
python-social-auth==0.2.12
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
qrcode==5.1
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tinycss2==0.6.1
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
psycopg2-binary
